24/02/2025

+++ src/Controller/NoteController.php
@@ -4,7 +4,6 @@ namespace App\Controller;
 
 use App\Entity\Note;
 use App\Form\NoteType;
-// use App\Repository\NoteRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -15,17 +14,21 @@ use DateTimeImmutable;
 use DateTimeZone;
 
 #[Route('/note')]
+
 final class NoteController extends AbstractController
+
 {
+    
     #[Route('/new', name: 'app_note_add', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager, SluggerInterface $slugger): Response
     {
+        if(!$this->isGranted('ROLE_USER')) return $this->redirectToRoute('app_home');
+        // virifie si l'utilisateur est connecté
         $note = new Note();
         $note->setUser($this->getUser());
         $note->setCreatedAt(new DateTimeImmutable('now', new DateTimeZone('Europe/Paris')));
         $form = $this->createForm(NoteType::class, $note);
         $form->handleRequest($request);
-
         if ($form->isSubmitted() && $form->isValid()) {
             $imageFile = $form->get('img')->getData();
             if ($imageFile) {
@@ -53,6 +56,7 @@ final class NoteController extends AbstractController
     #[Route('/{id}', name: 'app_note_show', methods: ['GET'])]
     public function show(Note $note): Response
     {
+        if($this->isGranted('ROLE_USER')) return $this->redirectToRoute('app_home');
         return $this->render('note/show.html.twig', [
             'note' => $note,
         ]);
@@ -61,6 +65,8 @@ final class NoteController extends AbstractController
     #[Route('/{id}/edit', name: 'app_note_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Note $note, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->isGranted('ROLE_USER')) return $this->redirectToRoute('app_home');
+
         $form = $this->createForm(NoteType::class, $note);
         $form->handleRequest($request);
 
@@ -79,11 +85,14 @@ final class NoteController extends AbstractController
     #[Route('/{id}', name: 'app_note_delete', methods: ['POST'])]
     public function delete(Request $request, Note $note, EntityManagerInterface $entityManager): Response
     {
+        if(!$this->isGranted('ROLE_USER')) return $this->redirectToRoute('app_home');
+
+
         if ($this->isCsrfTokenValid('delete' . $note->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($note);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('app_note_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_account', [], Response::HTTP_SEE_OTHER);
     }
 }
